refactor: converted sum distnaces tree to ts
